Refactor Startup finalizer
Refactor the startup finalizer by removing it's service component behavior
